requirements.txt
selenium~=4.8.2
undetected-chromedriver==3.0.3
webdriver-manager==3.5.4
Pillow~=9.4.0,
chromedriver-autoinstaller==0.4.0

setuptools~=60.2.0
future~=0.18.2
emoji~=2.2.0
Pillow~=9.4.0
mysql-connector-python~=8.0.32
WMI~=1.5.1
requests~=2.28.1
websockets~=10.4
pyperclip~=1.8.2
Flask~=2.2.3
psutil~=5.9.4
PySimpleGUI~=4.60.4